# api.yaml
openapi: 3.0.0
info:
  title: Ticket Management API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local server
paths:
...
  /movies:
    get:
      summary: List movies
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'

    post:
      summary: Create movie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieInput'
      responses:
        '201':
          description: Created
  /movies/{movieId}:
    get:
      summary: Get movie
      parameters:
        - in: path
          required: true
          name: movieId
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'

    put:
      summary: Replace movie
      parameters:
        - in: path
          required: true
          name: movieId
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieInput'
      responses:
        '200':
          description: Updated

    delete:
      summary: Delete movie
      parameters:
        - in: path
          required: true
          name: movieId
          schema: { type: string }
      responses:
        '204':
          description: No Content


  #############################
  # Sheet
  #############################
...
components:
  schemas:
...
    Movie:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        onair: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    MovieInput:
      type: object
      required: [title, onair]
      properties:
        title: { type: string }
        onair: { type: string, format: date-time }

    ########################################
    # Sheet
    ########################################
...